Change List<Account> To Account in AccountRepository, Fix typo in "PersonalDataReposioryImpl", add new update method to PersonalDataRepository
