+++ src/main/java/com/epam/aviasales/services/impl/AccountServiceImpl.java
@@ -40,7 +40,7 @@ public class AccountServiceImpl implements AccountService {
   }
 
   @Override
-  public List<Account> getAccountById(Long id) {
+  public Account getAccountById(Long id) {
     return accountRepository.getAccountById(id);
   }
